DequeTwoFIFOQeus
Using the First-In/First-Out (FIFO) method, queues organize their belongings. They function similarly to a pipe, with new objects being inserted at one end and old items being removed at the other. An enqueue operation is adding an item to one end of a queue. Queuing is the process of taking an item out of the opposite end.