FIFOQueueLinkedListArray
A data structure that keeps a list of elements in order is called a linked list. Every element in the list is referred to as a node, and each node has a reference to the list's subsequent node. The list's head and tail are the names given to the first and last nodes, respectively.